+++ test/runtests.jl
@@ -32997,6 +32997,7 @@ if test_group_enabled(:local_dependence) && RUN_RESEARCH_EVIDENCE_TESTS
 end
 if test_group_enabled(:generalized)
     include("model_family_contract.jl")
+    include("mgmfrm_validation_protocol.jl")
     include("generalized_guard_contract.jl")
     include("fixed_q_identification.jl")
 end